Done with export stats
